requirements.txt
dash
dash-bootstrap-components
pandas
openpyxl
python-dotenv
sqlalchemy
pyodbc
pydantic
